+++ app/src/main/java/it/pgp/xfiles/service/HTTPDownloadTask.java
@@ -51,7 +51,7 @@ public class HTTPDownloadTask extends RootHelperClientTask {
     }
 
     @Override
-    protected void onProgressUpdate(Pair<Integer,Integer>... values) {
+    protected void onProgressUpdate(Pair<Long,Long>... values) {
         ((MovingRibbon)mr).pb.setIndeterminate(false);
         mr.setProgress(values);
         builder.setProgress(100, (int) Math.round(values[0].i * 100.0 / values[0].j), false);
